+++ app/lib/desktop/pages/conversations/desktop_conversations_page.dart
@@ -5,6 +5,7 @@ import 'package:omi/pages/capture/widgets/widgets.dart';
 import 'package:omi/pages/conversations/widgets/processing_capture.dart';
 import 'package:omi/providers/capture_provider.dart';
 import 'package:omi/providers/conversation_provider.dart';
+import 'package:omi/utils/enums.dart';
 import 'package:omi/utils/responsive/responsive_helper.dart';
 import 'package:provider/provider.dart';
 import 'package:visibility_detector/visibility_detector.dart';
@@ -13,7 +14,7 @@ import 'widgets/desktop_conversation_card.dart';
 import 'widgets/desktop_empty_conversations.dart';
 import 'widgets/desktop_search_widget.dart';
 import 'widgets/desktop_search_result_header.dart';
-import 'widgets/desktop_premium_recording_widget.dart';
+import 'widgets/desktop_recording_widget.dart';
 
 /// Desktop conversations page - premium minimal design
 class DesktopConversationsPage extends StatefulWidget {
@@ -67,124 +68,150 @@ class _DesktopConversationsPageState extends State<DesktopConversationsPage>
   Widget build(BuildContext context) {
     super.build(context);
 
-    return Consumer<ConversationProvider>(
-      builder: (context, convoProvider, child) {
+    return Consumer2<ConversationProvider, CaptureProvider>(
+      builder: (context, convoProvider, captureProvider, child) {
+        final recordingState = captureProvider.recordingState;
+        final isRecording = recordingState == RecordingState.systemAudioRecord;
+        final isInitializing = recordingState == RecordingState.initialising;
+        final isRecordingOrInitializing = isRecording || isInitializing || captureProvider.isPaused;
+
         return Container(
           color: ResponsiveHelper.backgroundSecondary.withOpacity(0.85),
-          child: CustomScrollView(
-            slivers: [
-              const SliverToBoxAdapter(child: SizedBox(height: 20)),
+          child: Stack(
+            children: [
+              // Main conversations content
+              AnimatedOpacity(
+                opacity: isRecordingOrInitializing ? 0.0 : 1.0,
+                duration: const Duration(milliseconds: 400),
+                child: CustomScrollView(
+                  slivers: [
+                    const SliverToBoxAdapter(child: SizedBox(height: 20)),
 
-              // Hardware cards section with better spacing
-              SliverToBoxAdapter(
-                child: FadeTransition(
-                  opacity: _fadeAnimation,
-                  child: Container(
-                    padding: const EdgeInsets.symmetric(horizontal: 32),
-                    child: Column(
-                      children: [
-                        const SpeechProfileCardWidget(),
-                        const SizedBox(height: 12),
-                        const UpdateFirmwareCardWidget(),
-                        const SizedBox(height: 16),
-                        const DesktopPremiumRecordingWidget(),
-                      ],
-                    ),
-                  ),
-                ),
-              ),
-
-              // Search section with premium styling
-              SliverToBoxAdapter(
-                child: FadeTransition(
-                  opacity: _fadeAnimation,
-                  child: Container(
-                    padding: const EdgeInsets.fromLTRB(32, 24, 32, 16),
-                    child: const DesktopSearchWidget(),
-                  ),
-                ),
-              ),
-
-              const SliverToBoxAdapter(child: DesktopSearchResultHeader()),
-              getProcessingConversationsWidget(convoProvider.processingConversations),
-
-              // Main conversations content with premium design
-              if (convoProvider.groupedConversations.isEmpty && !convoProvider.isLoadingConversations)
-                SliverToBoxAdapter(
-                  child: FadeTransition(
-                    opacity: _fadeAnimation,
-                    child: Container(
-                      padding: const EdgeInsets.symmetric(vertical: 80),
-                      child: const Center(
-                        child: DesktopEmptyConversations(),
+                    // Hardware cards section with better spacing
+                    SliverToBoxAdapter(
+                      child: FadeTransition(
+                        opacity: _fadeAnimation,
+                        child: Container(
+                          padding: const EdgeInsets.symmetric(horizontal: 32),
+                          child: Column(
+                            children: [
+                              const SpeechProfileCardWidget(),
+                              const SizedBox(height: 12),
+                              const UpdateFirmwareCardWidget(),
+                              const SizedBox(height: 16),
+                              const DesktopPremiumRecordingWidget(),
+                            ],
+                          ),
+                        ),
                       ),
                     ),
-                  ),
-                )
-              else if (convoProvider.groupedConversations.isEmpty && convoProvider.isLoadingConversations)
-                SliverToBoxAdapter(
-                  child: Container(
-                    padding: const EdgeInsets.symmetric(vertical: 80),
-                    child: Center(
-                      child: CircularProgressIndicator(
-                        valueColor: AlwaysStoppedAnimation<Color>(ResponsiveHelper.purplePrimary),
-                        strokeWidth: 2,
+
+                    // Search section with premium styling
+                    SliverToBoxAdapter(
+                      child: FadeTransition(
+                        opacity: _fadeAnimation,
+                        child: Container(
+                          padding: const EdgeInsets.fromLTRB(32, 24, 32, 16),
+                          child: const DesktopSearchWidget(),
+                        ),
                       ),
                     ),
-                  ),
-                )
-              else
-                SliverPadding(
-                  padding: const EdgeInsets.symmetric(horizontal: 32),
-                  sliver: SliverList(
-                    delegate: SliverChildBuilderDelegate(
-                      childCount: convoProvider.groupedConversations.length + 1,
-                      (context, index) {
-                        if (index == convoProvider.groupedConversations.length) {
-                          return VisibilityDetector(
-                            key: const Key('desktop-conversations-load-more'),
-                            onVisibilityChanged: (visibilityInfo) {
-                              var provider = Provider.of<ConversationProvider>(context, listen: false);
-                              if (provider.previousQuery.isNotEmpty) {
-                                if (visibilityInfo.visibleFraction > 0 &&
-                                    !provider.isLoadingConversations &&
-                                    (provider.totalSearchPages > provider.currentSearchPage)) {
-                                  provider.searchMoreConversations();
-                                }
-                              } else {
-                                if (visibilityInfo.visibleFraction > 0 && !provider.isLoadingConversations) {
-                                  provider.getMoreConversationsFromServer();
-                                }
-                              }
-                            },
-                            child: Container(
-                              padding: const EdgeInsets.symmetric(vertical: 32),
-                              child: convoProvider.isLoadingConversations
-                                  ? Center(
-                                      child: CircularProgressIndicator(
-                                        valueColor: AlwaysStoppedAnimation<Color>(ResponsiveHelper.purplePrimary),
-                                        strokeWidth: 2,
-                                      ),
-                                    )
-                                  : const SizedBox(height: 20),
-                            ),
-                          );
-                        }
 
-                        // Conversation groups with premium spacing
-                        var date = convoProvider.groupedConversations.keys.elementAt(index);
-                        List<ServerConversation> conversationsForDate = convoProvider.groupedConversations[date]!;
+                    const SliverToBoxAdapter(child: DesktopSearchResultHeader()),
+                    getProcessingConversationsWidget(convoProvider.processingConversations),
 
-                        return FadeTransition(
+                    // Main conversations content with premium design
+                    if (convoProvider.groupedConversations.isEmpty && !convoProvider.isLoadingConversations)
+                      SliverToBoxAdapter(
+                        child: FadeTransition(
                           opacity: _fadeAnimation,
-                          child: _buildConversationGroup(date, conversationsForDate, index == 0),
-                        );
-                      },
-                    ),
+                          child: Container(
+                            padding: const EdgeInsets.symmetric(vertical: 80),
+                            child: const Center(
+                              child: DesktopEmptyConversations(),
+                            ),
+                          ),
+                        ),
+                      )
+                    else if (convoProvider.groupedConversations.isEmpty && convoProvider.isLoadingConversations)
+                      SliverToBoxAdapter(
+                        child: Container(
+                          padding: const EdgeInsets.symmetric(vertical: 80),
+                          child: Center(
+                            child: CircularProgressIndicator(
+                              valueColor: AlwaysStoppedAnimation<Color>(ResponsiveHelper.purplePrimary),
+                              strokeWidth: 2,
+                            ),
+                          ),
+                        ),
+                      )
+                    else
+                      SliverPadding(
+                        padding: const EdgeInsets.symmetric(horizontal: 32),
+                        sliver: SliverList(
+                          delegate: SliverChildBuilderDelegate(
+                            childCount: convoProvider.groupedConversations.length + 1,
+                            (context, index) {
+                              if (index == convoProvider.groupedConversations.length) {
+                                return VisibilityDetector(
+                                  key: const Key('desktop-conversations-load-more'),
+                                  onVisibilityChanged: (visibilityInfo) {
+                                    var provider = Provider.of<ConversationProvider>(context, listen: false);
+                                    if (provider.previousQuery.isNotEmpty) {
+                                      if (visibilityInfo.visibleFraction > 0 &&
+                                          !provider.isLoadingConversations &&
+                                          (provider.totalSearchPages > provider.currentSearchPage)) {
+                                        provider.searchMoreConversations();
+                                      }
+                                    } else {
+                                      if (visibilityInfo.visibleFraction > 0 && !provider.isLoadingConversations) {
+                                        provider.getMoreConversationsFromServer();
+                                      }
+                                    }
+                                  },
+                                  child: Container(
+                                    padding: const EdgeInsets.symmetric(vertical: 32),
+                                    child: convoProvider.isLoadingConversations
+                                        ? Center(
+                                            child: CircularProgressIndicator(
+                                              valueColor: AlwaysStoppedAnimation<Color>(ResponsiveHelper.purplePrimary),
+                                              strokeWidth: 2,
+                                            ),
+                                          )
+                                        : const SizedBox(height: 20),
+                                  ),
+                                );
+                              }
+
+                              // Conversation groups with premium spacing
+                              var date = convoProvider.groupedConversations.keys.elementAt(index);
+                              List<ServerConversation> conversationsForDate = convoProvider.groupedConversations[date]!;
+
+                              return FadeTransition(
+                                opacity: _fadeAnimation,
+                                child: _buildConversationGroup(date, conversationsForDate, index == 0),
+                              );
+                            },
+                          ),
+                        ),
+                      ),
+
+                    const SliverToBoxAdapter(child: SizedBox(height: 80)),
+                  ],
+                ),
+              ),
+
+              // Full-screen recording overlay
+              if (isRecordingOrInitializing)
+                AnimatedOpacity(
+                  opacity: isRecordingOrInitializing ? 1.0 : 0.0,
+                  duration: const Duration(milliseconds: 400),
+                  child: Container(
+                    width: double.infinity,
+                    height: double.infinity,
+                    child: const DesktopPremiumRecordingWidget(),
                   ),
                 ),
-
-              const SliverToBoxAdapter(child: SizedBox(height: 80)),
             ],
           ),
         );
